(PC-10174) Render la page ResetPasswordEmailSent après le reset password
