+++ src/features/auth/forgottenPassword/ForgottenPassword/index.ts
@@ -1,0 +1,1 @@
+export * from './ForgottenPassword'
