+++ src/features/auth/forgottenPassword/ResetPasswordEmailSent/index.ts
@@ -1,0 +1,1 @@
+export * from './ResetPasswordEmailSent'
