+++ src/features/auth/forgottenPassword/ResetPasswordEmailSent/tests/ResetPasswordEmailSent.native.test.tsx
@@ -7,14 +7,19 @@ import waitForExpect from 'wait-for-expect'
 
 import { navigateToHome } from 'features/navigation/helpers'
 import { RootStackParamList } from 'features/navigation/RootNavigator'
-import { flushAllPromises, act, fireEvent, render } from 'tests/utils/web'
+import { flushAllPromises, act, fireEvent, render, cleanup } from 'tests/utils'
 
-import { ResetPasswordEmailSent } from './ResetPasswordEmailSent'
+import { ResetPasswordEmailSent } from '../ResetPasswordEmailSent'
+
+// eslint-disable-next-line local-rules/no-allow-console
+allowConsole({ error: true })
 
 jest.mock('@react-navigation/native', () => jest.requireActual('@react-navigation/native'))
 jest.mock('features/navigation/helpers')
 
 describe('<ResetPasswordEmailSent />', () => {
+  afterEach(cleanup)
+
   it('should match snapshot', async () => {
     const renderAPI = await renderInitialPage('ResetPasswordEmailSent')
     expect(renderAPI).toMatchSnapshot()
@@ -26,8 +31,8 @@ describe('<ResetPasswordEmailSent />', () => {
     await act(async () => {
       navigationRef.current?.navigate('ResetPasswordEmailSent')
     })
-    const leftIcon = renderAPI.getByTestId('leftIcon')
-    fireEvent.click(leftIcon)
+
+    fireEvent.press(renderAPI.getByTestId('leftIcon'))
 
     await waitForExpect(() => {
       expect(renderAPI.queryByText('PreviousScreenText')).toBeTruthy()
@@ -47,7 +52,7 @@ describe('<ResetPasswordEmailSent />', () => {
   it('should redirect to Home when clicking on Close icon', async () => {
     const renderAPI = await renderInitialPage('ResetPasswordEmailSent')
 
-    fireEvent.click(renderAPI.getByTestId('rightIcon'))
+    fireEvent.press(renderAPI.getByTestId('rightIcon'))
 
     await waitForExpect(() => {
       expect(navigateToHome).toBeCalled()
@@ -58,7 +63,7 @@ describe('<ResetPasswordEmailSent />', () => {
     const renderAPI = await renderInitialPage('ResetPasswordEmailSent')
 
     const checkEmailsButton = renderAPI.getByText('Consulter mes e-mails')
-    fireEvent.click(checkEmailsButton)
+    fireEvent.press(checkEmailsButton)
 
     await waitForExpect(() => {
       expect(openInbox).toHaveBeenCalled()
@@ -96,8 +101,6 @@ async function renderInitialPage(initialScreenName: keyof StackParams) {
       </TestStack.Navigator>
     </NavigationContainer>
   )
-  await act(async () => {
-    await flushAllPromises()
-  })
+  await act(flushAllPromises)
   return renderAPI
 }
