+++ src/libs/react-native-email-link.web.ts
@@ -1,0 +1,1 @@
+export const openInbox = () => console.log('openInbox web') // eslint-disable-line no-console
